refactor: moved <Insight /> and other related components in shared refactor: moved utils to shared/utils
