+++ src/shared/components/Insight/Chart.jsx
@@ -1,11 +1,11 @@
 import { Checkbox, TextButton } from '@dailykit/ui'
 import React, { useState } from 'react'
 import { Chart as GoogleChart } from 'react-google-charts'
-import '../../../../shared/styled/tableStyles.css'
+import '../../styled/tableStyles.css'
 import { useChart } from '../../hooks/useChart'
-import { Container } from '../Container'
-import { Dropdown, DropdownItem } from '../DropdownMenu'
-import { Flex } from '../../../../shared/components/Flex'
+import { Container } from './Container'
+import { Dropdown, DropdownItem } from './DropdownMenu'
+import { Flex } from '../Flex'
 
 export default function Chart({ chart, oldData, newData, isDiff }) {
    const [chartType, setChartType] = useState({ ...chart.config[0], index: 0 })
