+++ src/shared/components/Insight/generators/generateBarChartData.js
