+++ src/shared/components/Insight/generators/generateCalendarChartData.js
