+++ src/shared/components/Insight/generators/generatePieChartData.js
