+++ src/shared/components/Insight/index.jsx
@@ -5,9 +5,9 @@ import 'react-tabulator/lib/styles.css'
 import styled from 'styled-components'
 import { Flex, Toggle } from '@dailykit/ui'
 
-import '../../../../shared/styled/tableStyles.css'
+import '../../styled/tableStyles.css'
 import { useInsights } from '../../hooks/useInsights'
-import { Counter } from '../Counter'
+import { Counter } from './Counter'
 import Chart from './Chart'
 import Option from './Option'
 import { tableConfig } from './tableConfig'
