+++ src/shared/components/Insight/tableConfig.js
